Move find sync workers into namespace
