+++ app/services/find_sync/sync_provider_from_find.rb
@@ -24,7 +24,7 @@ module FindSync
           ),
         )
 
-        SyncCoursesFromFind.perform_async(provider.id, provider_recruitment_cycle_year)
+        FindSync::SyncCoursesFromFind.perform_async(provider.id, provider_recruitment_cycle_year)
       else
         @provider = create_or_update_provider(base_provider_attrs)
       end
